[EasyCore] Created CarbonImmutableNormalizer (#835)
* feature/PLU-3995-update-carbon-normalizer
- removed CarbonNormalizer
- created CarbonImmutableNormalizer

* feature/PLU-3995-update-carbon-normalizer
- updated tests
